quicklook-analytics/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
google-generativeai
google-cloud-storage
google-cloud-core
numpy
scikit-learn
